Split MetadataDatabasePlugin to a different class per language
